Implement services

+++ account/src/main/java/timeboard.account/ExcelTASReport.java
@@ -1,5 +1,31 @@
 package timeboard.account;
 
+/*-
+ * #%L
+ * account
+ * %%
+ * Copyright (C) 2019 Timeboard
+ * %%
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ * 
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ * 
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ * #L%
+ */
+
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
@@ -60,9 +86,9 @@ public class ExcelTASReport extends AbstractExcelReport {
                     .setCellValue(tasData.getBusinessCode());
 
             List<String> dayMonthName = tasData.getDayMonthNames();
-            Map<Integer, BigDecimal> mapWorkedDAys = tasData.getWorkedDays();
-            Map<Integer, BigDecimal> mapOffDays = tasData.getOffDays();
-            Map<Integer, BigDecimal> mapOtherDays = tasData.getOtherDays();
+            Map<Integer, Double> mapWorkedDAys = tasData.getWorkedDays();
+            Map<Integer, Double> mapOffDays = tasData.getOffDays();
+            Map<Integer, Double> mapOtherDays = tasData.getOtherDays();
             Map<Integer, String> mapComments = tasData.getComments();
 
             for (int i = 0; i < tasData.getDayMonthNames().size(); i++) {
